pyproject.toml
[project]
name = "labjack_module"
version = "2.1.0"
description = "Software for automatting a PF400 robot"
authors = [
    {name = "Ryan D. Lewis", email="ryan.lewis@anl.gov"},
    {name = "Doga Ozgulbas", email="dozgulbas@anl.gov"},
    {name = "Abe Stroka", email="astroka@anl.gov"},
    {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
]
dependencies = [
    "madsci-node-module>=0.4.3",
    "labjack-ljm",
    "madsci-client>=0.4.3",
    "madsci-common>=0.4.3",
]
requires-python = ">=3.9.1,<3.13"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
homepage = "https://github.com/AD-SDL/labjack_module"


[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ruff>=0.12.0",
]
